api/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
Pillow
uwsgi
django-cors-headers
django-filter
requests
django-import-export
drf-excel
